map floating search input
