--- packages/uni_app/lib/view/locations/widgets/icons.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-/// Flutter icons LocationIcons
-/// Copyright (C) 2022 by original authors @ fluttericon.com, fontello.com
-/// This font was generated by FlutterIcon.com, which is derived from Fontello.
-///
-/// To use this font, place it in your fonts/ directory and include the
-/// following in your pubspec.yaml
-///
-/// flutter:
-///   fonts:
-///    - family:  LocationIcons
-///      fonts:
-///       - asset: fonts/LocationIcons.ttf
-///
-///
-///
-class LocationIcons {
-  LocationIcons._();
-
-  static const _kFontFam = 'LocationIcons';
-
-  static const IconData bookOpenBlankVariant =
-      IconData(0xe800, fontFamily: _kFontFam);
-  static const IconData bottleSodaClassic =
-      IconData(0xe801, fontFamily: _kFontFam);
-  static const IconData cashMultiple = IconData(0xe802, fontFamily: _kFontFam);
-  static const IconData coffee = IconData(0xe803, fontFamily: _kFontFam);
-  static const IconData printer = IconData(0xe804, fontFamily: _kFontFam);
-}
