+++ packages/uni_app/lib/view/map/widgets/marker_popup.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/model/entities/location.dart';
 import 'package:uni/model/entities/location_group.dart';
-import 'package:uni/view/locations/widgets/faculty_map.dart';
 
 class LocationMarkerPopup extends StatelessWidget {
   const LocationMarkerPopup(
@@ -54,7 +53,7 @@ class Floor extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final fontColor = FacultyMap.getFontColor(context);
+    final fontColor = _getFontColor(context);
 
     final floorString = 0 <= floor && floor <= 9 // To maintain layout of popup
         ? ' $floor'
@@ -86,6 +85,13 @@ class Floor extends StatelessWidget {
         ),
       ],
     );
+  }
+
+  // TODO(thePeras): Duplicated code
+  Color _getFontColor(BuildContext context) {
+    return Theme.of(context).brightness == Brightness.light
+        ? Theme.of(context).colorScheme.primary
+        : Theme.of(context).colorScheme.tertiary;
   }
 }
